% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{String teilen}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{Ein character Vektor mit der Länge 1 Element.}

\item{split}{Aufgrund von was aufteilen.}
}
\value{
Einen character Vektor.
}
\description{
String teilen
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
